requirements.txt
gensim
nltk
matplotlib
seaborn
setuptools
numpy
pandas
ipykernel
scikit-learn
chromadb
torch
ollama
pillow
requests
open-clip-torch
google-generativeai
python-dotenv
langchain
langchain-google-genai